Add .travis.yml: Run test-suite on Travis CI (travis-ci.com)
I *hope* this is still a sane decision in early 2021... :-S

But the GitHub Actions documentation says the storage limit
for GitHub Free would be 500 MB, which would be not nearly enough
for a full virtualenv and all the AI models used. ...
(That would rather be around 5 GB?)

Let's hope & see what Travis CI has to offer!