tests: helper: Add capture_stdout_stderr() and ..._cstderr()
That's right, we're able to capture output from C's stderr, now,
which was not caught by Python's `with contextlib.redirect_stderr()`.

This will be used to catch/check/filter-out warnings from faiss,
which luckily "just" use C stderr for output, nothing fancy.
So we can "just" replace the global stderr FILE-pointer temporarily.
This is technically non-portable (as stderr is supposed to be a macro),
but it seems to work, even where it's internally a slightly different
variable name, e.g., on OS X. (See the in-source list of URLs
as reference.)