build-index: Load CLIP (and arcface) lazily
The former like in query-index, although we there are more in need of
the implied caching, for multiple query-index.Search instances.
We're currently using just a single build-index.Scanner instance
everywhere, so caching will not kick in. (yet?)

Still, this saves 3.7s of forced startup time on my system
(3.3s for loading CLIP and 0.37s for loading arcface), which is
half the wall-clock time of a dry run with artificially coded forced-load.
That means this time will actually be effectively saved when a
build-index run doesn't need to build any CLIP or face features,
and therefore won't ever load the CLIP model.

For the case of getting --help to run more quickly, it doesn't do
anything, though. :(