GUI: Add initial thumbnail support
This currently only adds 24x24 (?) images into the table view
on the Images tab. Too small to really be useful, but we don't
want them larger as to not fit even less table lines on the screen
at the same time. An icon view is planned for larger thumbnails,
instead, but not implemented, yet.

Also, the thumbnails are currently loaded eagerly, so this should
probably be marked as WIP; but it runs and works (just takes
a little time to process all the images returned by a search),
so I'll just keep this on its own branch.