GUI: WIP: Initial use of multiprocessing for thumbnail generation
Issue GH-3: "GUI: Provide thumbnails"

With multithreading, no additional CPU-bound performance
can be achieved with Python due to the Global Interpreter Lock (GIL):
<https://docs.python.org/3/glossary.html#term-global-interpreter-lock>

That is, we can well have multiple OS-level threads, but they can't run
python code at the same time. That is where multiprocessing kicks in:
It side-steps the GIL by using multiple OS-level processes. This,
of course, gives rise to new problems, like encapsulating/decapsulating
the data moved to/from these processes, or managing the processes
(especially on exit), or it may even be hard to start the processes
at all when fork() is used and there have already been multiple threads
started in the main process. (To combat this, the multiprocessing
start method is changed (on unix: from fork) to spawn. That is,
every child process will start its own python interpreter instead of
sharing some state with the parent process.)

So this is a Work-In-Progress (WIP) attempt to let at least the
large thumbnails be loaded/calculated in parallel. The code
has grown organically (regarding the reuse of existing structures),
and it seems the first 3 images will be loaded at once, then nothing
seems to happen, then all the other images appear at once. Maybe
this has something to do with missing or too seldom polling, or
I don't know...