GUI: Try to optimize out repetitive thumbnail requests
This *may* avoid a problem where, after using "more results" many times,
there's what appears to be a CPU hang (100% CPU on the main process
and 50% CPU on each worker process), but does in fact go away
when simply doing a new search (with non-command, non-empty search input).

Maybe this commit's change fixes the problem, at least I couldn't
reproduce it again.

In any case, this re-adds some filter-out-existing-requests logic
to the thumbnail background processing, as was present with the previous
ImagesWorker (QThread) and was simply forgotten during the change-over to
ThumbnailGenerator (multiprocessing).